+++ config.tmpl.php
@@ -6,3 +6,5 @@ $dbuser = "sbviatges";
 $dbpass = "sbviatges";
 $dbbase = "sbschedule";
 $dbport = "8889";
+
+$mtxtfooter = "\r\n\r\n"."This email was sent automatically by SB Schedule.\r\nSB Schedule is a product by Saskia Brückner Kreativagentur · World of SB";