+++ CHANGELOG.md
@@ -4,6 +4,15 @@
 
 | Mayor | Minor | Build |
 | ----- | ----- | ----- |
+| 0     | 3     | 0     |
+
+- Added "accept", "decline", "save", "cancel", "create" event
+- Added archived events list
+
+* * *
+
+| Mayor | Minor | Build |
+| ----- | ----- | ----- |
 | 0  | 2     | 0     |
 
 - Bugfix: Login with uppercase letters in email address failed.